feat(Jpa): Add entities and repositories

+++ src/main/java/stream/arepresas/cryptotracker/external/coinMarket/dto/CoinMarketApiStatus.java
@@ -7,8 +7,8 @@ import java.util.Date;
 @Data
 public class CoinMarketApiStatus {
   private Date timestamp;
-  private int error_code;
+  private Long error_code;
   private String error_message;
-  private int elapsed;
-  private int credit_count;
+  private Long elapsed;
+  private Long credit_count;
 }
